Patch submitted by Doug Bryant [06/Jul/04 11:42 PM]:
Here is a patch for the examples that include groovy examples for every
java example, including Mark's new JNDI and Serialized fibonacci examples.

I moved all the packages from the org.drools.examples.java.* to the
org.drools.examples package.

Now, in each example, petstore/state/etc, there are two drl files. A
petstore.java.drl and a petstore.groovy.drl (or whatever subject the
example is dealing with)

I modified all the classes with main methods to check for an argument to
main and assume that argument specifies the correct drl to run against.

I modified the maven.xml to run all the examples. You run the examples
now by appending -java or -groovy to the example you want to run.
For instance, fibonacci is now fibonacci-java or fibonacci-groovy.
This makes it easy to add python examples later on.

project.xml is updated to include groovy dependencies.