These files represent a new example: PrimeFactors. It demonstrates the dynamic runtime creation of a number of rules (170 to be exact), and the subsequent assertion of a large number of facts (1000 to be exact).
The asserted facts are the numbers from 1 to 1000 and the rules perform a prime
factorization of the numbers. For example, 819 is factored into [13, 7, 3, 3].

Try "maven primefactors-java" to compare the execution times of three different
ways of asserting all the facts. Try "primefactors-java-debug" to see a smaller
sample run that prints the actual factorized numbers in the output.